+++ iac/terraform/2_apps_deployment/main.tf
@@ -1,14 +1,13 @@
-resource "null_resource" "app_deployment" {
-  for_each = { for i, app in var.app_deployments : i => app if app.status == "OK" }
-
-  triggers = {
-    project_id = each.value.project_id
-    app_name   = each.value.app_name
+output "app_deployments_details" {
+  description = "Details for deploying applications, including project_id, app_name, and environment variables."
+  value = {
+    for i, app in var.app_deployments : i => {
+      project_id  = app.project_id
+      app_name    = app.app_name
+      environment = app.environment
+      working_dir = abspath("${path.module}/../../../applications/${app.app_name}")
+      start_script = abspath("${path.module}/../../../applications/${app.app_name}/start.sh")
+    } if app.status == "OK"
   }
-
-  provisioner "local-exec" {
-    command     = "./start.sh"
-    working_dir = abspath("${path.module}/../../../applications/${each.value.app_name}")
-    environment = each.value.environment
-  }
+  sensitive = true # Mark as sensitive to avoid printing secrets to the console
 }